Milestones 1 (Phase 2)/requirements.txt
streamlit
numpy
pandas
tensorflow-cpu